refactor assignmentTile
