+++ frontend/src/store/entities/assignments/types.ts
@@ -19,6 +19,7 @@ export type AssignmentsType = {
   review: string;
   assignment_root: number;
   is_public: boolean;
+  average_grade: string;
 };
 
 export type AssignmentsResponseType = {
