+++ frontend/src/utils/helperFunction/formatDate.ts
@@ -1,0 +1,5 @@
+export const formatDate = (dateString: string) => {
+    const options:  Intl.DateTimeFormatOptions = { year: "numeric", month: "short", day: "numeric" };
+    return new Date(dateString).toLocaleDateString("en-US", options);
+  };
+
